frontend/flask/requirements.txt
flask
uwsgi
redis

celery

requests
requests-cache

pyyaml

numpy
scipy
pandas
networkx
